[RENAME] #56 - 패키지명 이름 수정
- 이름 수정을 인한 커밋
