+++ src/main/java/com/core/linkup/club/club/repository/ClubCustomRepository.java
@@ -1,7 +1,7 @@
 package com.core.linkup.club.club.repository;
 
 import com.core.linkup.club.club.entity.Club;
-import com.core.linkup.club.club.requset.ClubSearchRequest;
+import com.core.linkup.club.club.request.ClubSearchRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
