+++ src/main/java/com/core/linkup/club/club/request/ClubUpdateRequest.java
@@ -1,12 +1,10 @@
-package com.core.linkup.club.club.requset;
+package com.core.linkup.club.club.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
-import java.util.List;
-
 @Builder
-public record ClubCreateRequest(
+public record ClubUpdateRequest (
         String title,
         String introduction,
         @JsonProperty("category")
@@ -15,7 +13,6 @@ public record ClubCreateRequest(
         Boolean clubAccessibility,
         String detailedIntroduction,
         String clubThumbnail,
-        String applicationIntroduction,
-        List<ClubQuestionRequest> clubQuestions
+        String applicationIntroduction
 ) {
 }
