+++ src/main/java/com/core/linkup/club/club/service/ClubService.java
@@ -10,11 +10,11 @@ import com.core.linkup.club.club.repository.ClubMemberRepository;
 import com.core.linkup.club.club.repository.ClubRepository;
 import com.core.linkup.club.club.response.ClubApplicationResponse;
 import com.core.linkup.club.club.response.ClubSearchResponse;
-import com.core.linkup.club.club.requset.ClubSearchRequest;
+import com.core.linkup.club.club.request.ClubSearchRequest;
 import com.core.linkup.club.club.repository.ClubQuestionRepository;
-import com.core.linkup.club.club.requset.ClubApplicationRequest;
-import com.core.linkup.club.club.requset.ClubCreateRequest;
-import com.core.linkup.club.club.requset.ClubUpdateRequest;
+import com.core.linkup.club.club.request.ClubApplicationRequest;
+import com.core.linkup.club.club.request.ClubCreateRequest;
+import com.core.linkup.club.club.request.ClubUpdateRequest;
 import com.core.linkup.common.exception.BaseException;
 import com.core.linkup.common.response.BaseResponseStatus;
 import com.core.linkup.member.entity.Member;
